refactor(seo): move detail-page JSON-LD builders into jsonLdSchemas
The three detail-page loaders hand-built their schema.org JSON-LD inline;
the publications loader alone was 234 lines of @type mapping and field
assembly. Extract pure builders into jsonLdSchemas.ts:

- buildPublicationJsonLd(publication, base)
- buildCommunicationJsonLd(communication, base)
- buildDhProjectJsonLd(project, base)

Logic moved verbatim (identical property order → identical serialised
output); base path is passed in so the builders stay free of framework
runtime imports and are unit-testable. The loaders are now thin lookups
(publications/[id]/+page.ts: 234 -> 21 lines).
