+++ src/routes/communications/[id]/+page.ts
@@ -1,10 +1,8 @@
 import { error } from '@sveltejs/kit';
 import { base } from '$app/paths';
 import { allCommunications } from '$lib/data/communications/index';
+import { buildCommunicationJsonLd } from '$lib/utils/jsonLdSchemas';
 import type { PageLoad } from './$types';
-import type { EventJsonLd } from '$lib/types/jsonld';
-import { formatAuthors, formatJsonLdDate } from '$lib/types/jsonld';
-import { website } from '$lib/data/siteConfig';
 
 export const load: PageLoad = ({ params }) => {
 	const communication = allCommunications.find((comm) => comm.id === params.id);
@@ -13,69 +11,7 @@ export const load: PageLoad = ({ params }) => {
 		throw error(404, 'Communication not found');
 	}
 
-	// Build JSON-LD structured data using Event schema
-	// Academic presentations/talks are modeled as Events with performer
-	const jsonLdObject: Partial<EventJsonLd> = {
-		'@context': 'https://schema.org',
-		'@type': 'Event',
-		name: communication.title,
-		description: communication.abstract,
-		url: `${base}/communications/${communication.id}`
-	};
-
-	// Date
-	if (communication.dateISO) {
-		jsonLdObject.startDate = formatJsonLdDate(communication.dateISO);
-	}
-
-	// Location
-	if (communication.location || communication.country) {
-		const locationParts = [communication.location, communication.country]
-			.filter(Boolean)
-			.join(', ');
-		jsonLdObject.location = {
-			'@type': 'Place',
-			name: locationParts
-		};
-	}
-
-	// Organizer (conference name)
-	if (communication.conference) {
-		jsonLdObject.organizer = {
-			'@type': 'Organization',
-			name: communication.conference
-		};
-	}
-
-	// Performers (authors/presenters)
-	if (communication.authors && communication.authors.length > 0) {
-		const authors = formatAuthors(communication.authors);
-		// Add URL to the primary author
-		jsonLdObject.performer = authors.map((a) =>
-			a.name === 'Frédérick Madore' ? { ...a, url: website.url } : a
-		);
-	}
-
-	// Image
-	if (communication.heroImage?.src) {
-		jsonLdObject.image = `${base}/${communication.heroImage.src}`;
-	} else if (communication.image) {
-		jsonLdObject.image = `${base}/${communication.image}`;
-	}
-
-	// Keywords
-	if (communication.tags) {
-		jsonLdObject.keywords = communication.tags.join(', ');
-	}
-
-	// Language
-	if (communication.language) {
-		jsonLdObject.inLanguage = Array.isArray(communication.language)
-			? communication.language[0]
-			: communication.language;
-	}
-
-	const jsonLdString = JSON.stringify(jsonLdObject);
+	const jsonLdString = JSON.stringify(buildCommunicationJsonLd(communication, base));
 
 	return {
 		communication,
