+++ src/routes/digital-humanities/[id]/+page.ts
@@ -3,9 +3,7 @@ import { base } from '$app/paths';
 import type { PageLoad, PageLoadEvent } from './$types';
 import { allDhProjects } from '$lib/data/digital-humanities';
 import type { DigitalHumanitiesProject } from '$lib/types/digitalHumanities';
-import type { CreativeWorkJsonLd } from '$lib/types/jsonld';
-import { formatAuthor } from '$lib/types/jsonld';
-import { website } from '$lib/data/siteConfig';
+import { buildDhProjectJsonLd } from '$lib/utils/jsonLdSchemas';
 
 export const load: PageLoad = (event: PageLoadEvent) => {
 	const project: DigitalHumanitiesProject | undefined = allDhProjects.find(
@@ -16,32 +14,7 @@ export const load: PageLoad = (event: PageLoadEvent) => {
 		throw error(404, 'Project not found');
 	}
 
-	// Build JSON-LD structured data using CreativeWork schema
-	const jsonLdObject: Partial<CreativeWorkJsonLd> = {
-		'@context': 'https://schema.org',
-		'@type': project.linkUrl ? 'WebSite' : 'CreativeWork',
-		name: project.title,
-		description: project.seoDescription || project.shortDescription,
-		url: project.linkUrl || `${base}/digital-humanities/${project.id}`
-	};
-
-	// Author
-	const author = formatAuthor('Frédérick Madore');
-	jsonLdObject.author = [{ ...author, url: website.url }];
-
-	// Image
-	if (project.heroImageUrl || project.imageUrl) {
-		jsonLdObject.image = `${base}/${project.heroImageUrl || project.imageUrl}`;
-	}
-
-	// Keywords
-	if (project.seoKeywords && project.seoKeywords.length > 0) {
-		jsonLdObject.keywords = project.seoKeywords.join(', ');
-	} else if (project.skills && project.skills.length > 0) {
-		jsonLdObject.keywords = project.skills.join(', ');
-	}
-
-	const jsonLdString = JSON.stringify(jsonLdObject);
+	const jsonLdString = JSON.stringify(buildDhProjectJsonLd(project, base));
 
 	return {
 		project,
